cleaned up the code a bit and emptied app by assing MainComponent instead
