+++ README.md
@@ -1,6 +1,6 @@
 <h1 align="center">
   <a href="">
-    <img src="/src/assets/happy-thoughts.svg" alt="Project Banner Image">
+    <img src="/src/assets/examples/finished-example.png" alt="Project Banner Image">
   </a>
 </h1>
 
@@ -8,17 +8,6 @@
 
 In this week's project, you'll be able to practice your React state skills by fetching and posting data to an API.
 
-## Getting Started with the Project
-
-### Dependency Installation & Startup Development Server
-
-Once cloned, navigate to the project's root directory and this project uses npm (Node Package Manager) to manage its dependencies.
-
-The command below is a combination of installing dependencies, opening up the project on VS Code and it will run a development server on your terminal.
-
-```bash
-npm i && code . && npm run dev
-```
 
 ### The Problem
 
@@ -26,10 +15,4 @@ Describe how you approached to problem, and what tools and techniques you used t
 
 ### View it live
 
-Every project should be deployed somewhere. Be sure to include the link to the deployed project so that the viewer can click around and see what it's all about.
-
-## Instructions
-
-<a href="instructions.md">
-   See instructions of this project
-  </a>
+https://happy-machine.netlify.app/
