+++ src/App.css
@@ -34,13 +34,15 @@ h2 {
   display: grid;
   grid-template-columns: 1fr 1fr;
   grid-template-rows: 1fr;
+  word-wrap: break-word;
 }
 
-#unique-message {
+.unique-message {
   grid-column: 1 / 3;
   grid-row: 1;
   padding-bottom: 1em;
-  display: flex;
+  display: inline-block;
+  word-wrap: break-word;
 }
 
 #likes {
