+++ src/components/LikeThought.jsx
@@ -1,15 +1,17 @@
 import { useState } from 'react'
 import PropTypes from 'prop-types'
 
+//Adds a like when you click on the heart button, and post it to the API
 export const LikeThought = ({ message, index, setMessageUpdate, messageUpdate }) => {
 
   const [like, setLike] = useState(message.hearts)
 
+  //Listening to the button and adds 1 like
   const handleClick = async (event) => {
     event.preventDefault();
-    // Use the functional update pattern
     setLike((prevHearts) => prevHearts + 1);
   
+    //Posts the like the API if the request is successfull and sets the messageUpdate to true to trigger a rerender
     try {
       const response = await fetch(`https://happy-thoughts-ux7hkzgmwa-uc.a.run.app/thoughts/${message._id}/like`, {
         method: 'POST',
