--- todo.md
@@ -1,4 +1,0 @@
-TO DO:
-
-- Fixa input placeholder, top left
-- move APP functions into child component
